Added AuthService with login method & a new exception with this.

+++ src/main/java/com/beachg/backend/dtos/auth/AuthResponse.java
@@ -2,6 +2,6 @@ package com.beachg.backend.dtos.auth;
 
 public record AuthResponse(
         String token,
-        int clientId
+        Long clientId
 ) {
 }
